added token authentification, added user mongoose schema, added user and login controllers, correct jest blog tests, added middlewares

+++ controllers/blogs.js
@@ -1,29 +1,45 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const userExtractor = require('../utils/middleware').userExtractor
+
 
 //no try catch since we are using express-async-errors
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body
+
+  const user = request.user
 
   const newBlog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
+    user: user._id
   })
 
   const blog = await newBlog.save()
+  user.blogs = user.blogs.concat(blog._id)
+  await user.save()
   response.status(201).json(blog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+
+  if(blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }else{
+    return response.status(401).json({error: 'token does not match'})
+  }
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
